pypi/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "miragepy"
version = "0.0.2"
authors = [
    { name = "Kaan Giray Buzluk", email = "kaangiray26@protonmail.com" },
]
description = "🏖 mirage is a static site generator in Python"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent"
]
dependencies = [
    "quart",
    "markdown",
    "pyyaml"
]

[project.scripts]
mirage = "mirage.mirage:Mirage"

[project.urls]
"Homepage" = "https://github.com/kaangiray26/mirage"
"Bug Tracker" = "https://github.com/kaangiray26/mirage/issues"